requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.2
fastapi==0.99.1
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.10
python-dotenv==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
